saving rating in db

+++ setup.py
@@ -1,7 +1,7 @@
 from flask import Flask, flash, render_template, request, jsonify, redirect, session, url_for, send_file
 from flask_sqlalchemy import SQLAlchemy
 import os
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Float
 
 from werkzeug.datastructures import ImmutableMultiDict
 from speech_rating_app import rate
@@ -10,7 +10,7 @@ import psycopg2
 ALLOWED_EXTENSIONS = {'wav'}
 setup = Flask(__name__)
 
-# app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://postgres:5432@localhost/xise'
+# setup.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://postgres:5432@localhost/xise'
 setup.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 setup.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
 db = SQLAlchemy(setup)
@@ -20,9 +20,11 @@ class Employee(db.Model):
     __tablename__ = 'employee'
     id = Column(Integer, primary_key=True)
     speech_file = Column(String(200))
+    rating = Column(Float)
 
-    def __init__(self, speech_file):
+    def __init__(self, speech_file, rating):
         self.speech_file = speech_file
+        self.rating = rating
 
 
 def allowed_file(filename):
@@ -35,11 +37,13 @@ def audio_rating(id):
     id = request.view_args['id']
 
     employee = db.session.query(Employee).filter_by(id=id).first()
-    # file = download(employee.speech_file)
-
-    rating = rate(employee.speech_file)
-    os.remove(employee.speech_file)
-    return jsonify({"rating": rating}), 200
+    if employee.rating == -1:
+        rating = rate(employee.speech_file)
+        os.remove(employee.speech_file)
+        employee.rating = rating
+        db.session.commit()
+        return jsonify({"rating": rating}), 200
+    return jsonify({"rating": employee.rating}), 200
 
 
 @ setup.route('/speech/audio/upload', methods=['POST'])
@@ -47,21 +51,12 @@ def audio_upload():
     if 'file' not in request.files:
         return jsonify('No file present'), 400
     f = request.files['file']
-
-    if f.filename == '':
-        return jsonify('No selected file'), 400
-
     if allowed_file(f.filename):
-        # filename = upload(file)
-        # open("files/"+f.filename, 'w').write(f)
         f.save(f.filename)
-        employee = Employee(f.filename)
+        employee = Employee(f.filename, -1)
         db.session.add(employee)
         db.session.commit()
-        inserted_id = employee.id
-        print(employee.speech_file)
-
-        return jsonify({"id": inserted_id}), 201
+        return jsonify({"id": employee.id}), 201
     else:
         return jsonify("No wav file found"), 400
